+++ samples/serverless/main-app/Activities.cs
@@ -6,7 +6,6 @@ namespace Microsoft.DurableTask.Samples.Serverless.MainApp;
 internal static class ServerlessTaskNames
 {
     public const string LocalHello = "LocalHello";
-    public const string RemoteHello = "RemoteHello";
     public const string HelloOrchestrator = nameof(HelloOrchestrator);
 }